+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/basic1/Basic1Activity.java
@@ -14,26 +14,23 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.basic2;
+package im.ene.lab.toro.sample.presentation.basic1;
 
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-import im.ene.lab.toro.sample.R;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
 
 /**
- * Created by eneim on 6/29/16.
- *
- * Sample ViewHolder which holds no Video. Or it doesn't request support from Toro.
+ * Created by eneim on 6/30/16.
  */
-public class Basic2NormalViewHolder extends Basic2ViewHolder {
+public class Basic1Activity extends AppCompatActivity {
 
-  public static final int LAYOUT_RES = R.layout.vh_normal_view;
-
-  public Basic2NormalViewHolder(View itemView) {
-    super(itemView);
-  }
-
-  @Override public void bind(RecyclerView.Adapter adapter, Object item) {
-
+  @Override protected void onCreate(@Nullable Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    if (savedInstanceState == null) {
+      getSupportFragmentManager().beginTransaction()
+          .replace(android.R.id.content, Basic1ListFragment.newInstance())
+          .commit();
+    }
   }
 }
