+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/basic2/Basic2Adapter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.basic2;
+package im.ene.lab.toro.sample.presentation.basic2;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
