+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/basic2/Basic2ViewHolder.java
@@ -14,26 +14,25 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.basic3;
+package im.ene.lab.toro.sample.presentation.basic2;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
-import im.ene.lab.toro.sample.R;
 
 /**
  * Created by eneim on 6/29/16.
  *
- * Sample ViewHolder which holds no Video. Or it doesn't request support from Toro.
+ * A sample {@link RecyclerView.ViewHolder} skeleton.
  */
-public class Basic3NormalViewHolder extends Basic3ViewHolder {
+public abstract class Basic2ViewHolder extends RecyclerView.ViewHolder {
 
-  public static final int LAYOUT_RES = R.layout.vh_normal_view;
+  static int TYPE_VIDEO = 1;
 
-  public Basic3NormalViewHolder(View itemView) {
+  static int TYPE_NORMAL = 2;
+
+  public Basic2ViewHolder(View itemView) {
     super(itemView);
   }
 
-  @Override public void bind(RecyclerView.Adapter adapter, Object item) {
-
-  }
+  public abstract void bind(RecyclerView.Adapter adapter, Object item);
 }
