+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/basic3/Basic3NormalViewHolder.java
@@ -14,23 +14,26 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.basic2;
+package im.ene.lab.toro.sample.presentation.basic3;
 
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+import im.ene.lab.toro.sample.R;
 
 /**
- * Created by eneim on 6/30/16.
+ * Created by eneim on 6/29/16.
+ *
+ * Sample ViewHolder which holds no Video. Or it doesn't request support from Toro.
  */
-public class Basic2Activity extends AppCompatActivity {
+public class Basic3NormalViewHolder extends Basic3ViewHolder {
 
-  @Override protected void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    if (savedInstanceState == null) {
-      getSupportFragmentManager().beginTransaction()
-          .replace(android.R.id.content, Basic2ListFragment.newInstance())
-          .commit();
-    }
+  public static final int LAYOUT_RES = R.layout.vh_normal_view;
+
+  public Basic3NormalViewHolder(View itemView) {
+    super(itemView);
+  }
+
+  @Override public void bind(RecyclerView.Adapter adapter, Object item) {
+
   }
 }
